requirements.txt
fastapi
uvicorn[standart]
psycopg2-binary
